MM-24533: fix channel name generation (#65)
If the channel name consisted of solely invalid characters, they were all stripped out, the server would reject the too short name, and then we'd add a `-<random string>`. This, itself, would then be invalid, and the process would be completely indefinitely, leaving the plugin in an infinite loop.

Fix this by eliminating the loop and instead trying at most once. Also reject channel names that are too short up front for simplicity.

Fixes: https://mattermost.atlassian.net/browse/MM-24533