MM-25396: look for channel export in redux
Instead of checking `window.plugins`, look for the channel export in the Redux store instead. My original idea of using `window.plugins` was flawed because the webapp doesn't actually /remove/ plugins from that array on disable.

Using Redux wasn't as straightforward as I hoped. As part of https://mattermost.atlassian.net/browse/MM-17087, the `RECEIVED_WEBAPP_PLUGIN` action is no longer emitted when a plugin is enabled, meaning the Redux store wasn't kept up-to-date for /enabling/ a plugin after load. For now, I've fixed this in a linked channel export plugin PR to emit the event manually, but I'll circle back and fix this in the webapp proper.

Fixes: https://mattermost.atlassian.net/browse/MM-25396