Make GetIncidents return consistent
If the list of incidents returned is empty, return always a nil value,
not an empty list.

This is needed because the incident list is declared uninitialized and,
when the returned list from the SQL query is populated but the paging
ends up filtering out all incidents, an empty list is returned instead.