MM-28034: fix RHS opener
Maintain in Redux a list of all known incident channels for the current user, lazy loaded on start, team changes, and whenever a websocket event occurs signalling the start of an incident.

Leverage this list to unambiguously decide when a channel is an
incident, and tidy up the RHS opener logic to open the RHS reliably.

Fixes: https://mattermost.atlassian.net/browse/MM-28034