prompt for incident name and description
When starting an incident, prompt for an incident name (vs. the `channel name` currently provided), and then add a prompt for an optional description. The description ends up as the channel header.

When starting an incident from a post, this description will be pre-filled with the markdown we currently use to record a link to the post, but can be modified freely. We only show `The channel was created by the Incident Response plugin.` in the channel header if no description is provided.

Fixes: https://mattermost.atlassian.net/browse/MM-28311