MM-26166 incident metadata endpoint (#323)
* MM-26166: /metadata vs. /details

Switch the existing `/incidents/<id>/details` endpoint to `/incidents/<id>/metadata`, returning /only/ the metadata and not the incident as well, eliminating the need for an `IncidentWithDetails` structure and corresponding juggle.

This change is primarily about v1.0 API stability.

Fixes: https://mattermost.atlassian.net/browse/MM-26166

* unify /incidents and /incidents/<id> responses

Always return the checklist contents. There might be a performance overhead, but it should be minimal, and simplifies the developer API surface.