MM-26182: Standardize filtering parameters (#340)
All collection endpoints that support sort, direction and search_term (namely,
GET /incidents and GET /incidents/channels) are now consistent. However, there
are collections endpoints that only support a subset of those parameters. I did
not add any code to support the missing parameters because we don't need it
right now. The endpoints that don't support everything are:

- GET /playbooks supports sort and direction, but does not support search_term.
- GET /incidents/commanders does not support sort, order nor search_term.

I also removed the status value for the sort parameter because it feels
redundant:

- status filtering is usually done through the status parameter
- we also have the is_active value for the sort parameter, in case we wanna sort
  it this way