MM-29115: Update OpenAPI docs (#365)
This commit updates the API docs to match the current implementation of the API.
There are a bunch of changes, but in general they come down to the following:
- Reviewed existing endpoints:
  - Added missing parameters (query, path or body)
  - Added descriptions and examples to all of them
  - Make examples consistent so that the curl request does not show a completely
    different thing than the automatically generated examples.
  - Reviewed the responses and made them consistent.
  - In general, tried to make everything consistent.
- Added missing endpoints; namely:
  - /incidents/<id>: PATCH
  - /incidents/<id>/end: POST 
  - /incidents/<id>/restart: PUT
  - /incidents/<id>/next-stage-dialog: POST
  - /incidents/checklists/<id>/item/<id>/assignee: PUT
  - /incidents/checklists/<id>/item/<id>/run: POST
  - /incidents/checklist-autocomplete: GET

This is not perfect yet, as we can still improve the way the readers will skim
through this. We need a generic introduction, a more thorough description on the
authorization methods and a better organization of the left-hand sidebar. Of
course, we still need to host this somewhere. But, at least, the API is synced
(at least for today).