MM-29867: Commands to generate test data (#358)
This commit adds two new slash commands:
- `/incident test create-incident`, which accepts a playbook ID (that can be
chosen from the playbooks the user is a member of, using the autocomplete
system), a timestamp and an incident name. It creates an ongoing incident whose
creation date is set to the specified timestamp.
- `/incident test bulk-data`, which accepts a number of ongoing incidents, a
number of ended incidents, a begin and an end date, and an optional seed. It
creates as many ongoing and ended incidents as specified, all of them with their
creation date randomly picked between the begin and end dates. The seed, if
available, is used to get reproducible results. The names of the incidents are
randomly chosen from a list of incident names and a list of fake company names I
defined in the code.

Other things I did while I was here:
- I fixed the `sqlstore.NukeDB` function, which failed with a foreign key error
when trying to delete `IR_Playbook` before `IR_PlaybookMember`.
- I moved the `st` command as a subcommand of `/incident test`, for the sake of
consistency.
- I excluded `G404: Insecure random number source (rand)` rule from
[gosec](https://github.com/securego/gosec): it was complaining because I am
using `math/rand` instead of `crpyto/rand` to generate the random numbers needed
in the bulk data generation. Of course, we do not need `crypto/rand` for such a
task. And, if we ever need to do something crypto-related, I think we will know
by ourselves that we need to use `crypto/rand`.