MM-31243: broadcast channel tweaks
Keep showing `Unknown Channel` when a private channel is configured as the broadcast channel in the backstage, but restore showing `Select a channel` when none is selected at all.

As part of writing the unit tests here, I discovered it was impossible to type a channel name: search was broken. I restored this behaviour with a simple `includes` scan on id, name and display name.

Fixes: https://mattermost.atlassian.net/browse/MM-31243