MM-31242: fix unusual backstage shadows (#415)
Safari's treatment of `linear-gradient` has `transparent` meaning black
transparent (namely `rgba(0,0,0,0)`).
([Reference](https://stackoverflow.com/questions/30673204/css-linear-gradient-transparency-misbehaving-only-in-safari).
This is actually according to spec, but results in the undesirable
behaviour below as we try to fade out the image assets in the backstage:

Fix this by making a transparent version of the background colour over
which to apply the gradient: