MM-32387: track license && conditionally activate
Incident Collaboration is installed by default on E0 and E10, self-managed instances, but currently fails to activate with an `ERROR` posted in the logs. Furthermore, once a license is installed, the plugin does not start until the next server restart.

Change this behaviour to allow the plugin to activate, but reject API calls and hide all client-side behaviour. Listen for a license change, and automatically re-enable the plugin in existing client sessions.

Fixes: https://mattermost.atlassian.net/browse/MM-32387