MM-32387: activate but functionally disable without license (#469)
* fix channel admin promotion tests

My recent recent commit in #458 silently conflicted with some of the improved testing from #459.

* internalize pluginapi's store support

We need to remove the external enterprise check to apply our own checks that still allow for database initialization until an E20 license is installed.

* MM-32387: track license && conditionally activate

Incident Collaboration is installed by default on E0 and E10, self-managed instances, but currently fails to activate with an `ERROR` posted in the logs. Furthermore, once a license is installed, the plugin does not start until the next server restart.

Change this behaviour to allow the plugin to activate, but reject API calls and hide all client-side behaviour. Listen for a license change, and automatically re-enable the plugin in existing client sessions.

Fixes: https://mattermost.atlassian.net/browse/MM-32387

* removed unused use of retI

* Revert "internalize pluginapi's store support"

This reverts commit 02519a553adb4645ea57ab03946cef2d8ebdb975.

* depend on updated https://github.com/mattermost/mattermost-plugin-api

* go mod tidy