MM-31561: Commander automation (#483)
This commit continues the work on the automation features, adding the ability to
assign a user as a commander when an incident is created.

Changes in the DB:
- New columns in `IR_Playbook`:
  - `DefaultCommanderID`, which, unsurprisingly, contains the ID of the
    commander.
  - `DefaultCommanderEnabled`, that stores the on/off state of the setting.
- New column in `IR_Incident`:
  - `DefaultCommanderID`, which stores a copy of the data in the playbook's
    column when the incident is created. If the playbook's
    `DefaultCommanderEnabled` contains `false`, the data is not copied, so
    there's no need to add a `DefaultCommanderEnabled` column to the
    `IR_Incident` table.

Changes in the service: When the incident is created, the user selected is
assigned as commander (if the user is not invited, it is automatically invited
to the channel). I changed the messages from the bot to adapt to the cases where
the commander and reporter are the same or different.

This also adds a refactor of the permissions package, which no longer depends on
the incident package.