MM-34665: Pricing differentiation plan: backend support (#525)
This is the first of a series of PRs for supporting the pricing differentiation
plan features. This one actually does not change the behaviour of the product at
all.

It just adds a couple of new things, that are not yet used: a
GetNumPlaybooksForTeam method to the playbook service and a new /playbooks/count
endpoint, as well as support for new websocket events and a store entry that
maintain the number of playbooks created per team. These will support the limits
for E0/E10 users in the coming PRs.

The only important change moves the license-check middleware to every API
subrouter. This also adds a "feature flag", which is a constant set to false for
now, that controls the restriction of (for now) the playbook subrouter. As the
feature flag is false, this subrouter is still restricted to E20 users.

There's also a minor change in the code of the backstage navbar to remove the
Switch component, which was not needed.

List of changes:

* Move E20Middleware to every API subrouter

* Add GetNumPlaybooksForTeam method

* Remove Switch from backstage navbar

* Convert isE20LicensedOrDevelopment into a selector

* Add a new /playbooks/count API endpoint

* Track number of playbooks per team in the store

* Add websocket to track playbook create/delete

* Run make mocks

* Adapt tests

* Add E20 middleware to the /settings route

* Move websocket broadcast to playbook service