s/commander/owner/
In preparing this change, I first considered just changing the user facing text. This turned out to be extremely complicated to do correctly, with the words "commander" and "owner" often appearing in the same line of code, in strings, but having different intents. This also had the downside of leaving the API with the word "commander" instead of "owner".

This change replaces all instances of commander, except:
* database column names and values: I opted to preserve backwards compatibility here.
* telemetry properties: Avoiding churn on the looker side here.

One other change is included: namely the use of `Role changes` instead of `Commander changes` or `Owner changes`, as per the reviewed terminology spec. I've commented everywhere the changes aren't just a simple string replacement.