collapse into app package
Collapse the `permissions`, `incident`, and `playbook` packages into a monolithic `app` package. The `permissions` and `playbook` packages were already small, and this paves the way to release the words `incident` and `playbook` for use as variable names without shadowing the respective package.

Medium term, the plan will also be to remove the interfaces between the sqlstore, app and api packages and just consume each other directly.  This would result in all unit tests hitting the database, eliminating most mocking.