MM-39126 Use natural lang for default update time
When specifying the default update timer, we currently
have to use a modal with hour/minute/day for custom
duration. Instead, let's re-use the natural language
input from the status update modal. This supports
entering durations such as "2 minutes", "6 days",
"3 hours 5 seconds".

Replaces a lot of custom-written code with a call to
DateTimeInput.