Reorganize E2E Tests
Organize into the following directory structure, rendering the term
`backstage` obsolete in this context and treating `channels` as an
integration and not the `frontstage`.
  - cypress/integration (all specs live here)
    - playbooks/
    - runs/
    - channels/
      - rhs/
      - slash_command/
    - (other global tests)

The top level `describe` of each spec now contains the path (e.g.  `channels > rhs > spec_name`). There's probably still room for improvement here, both in the organization and naming, but this gets us past the hump as it were.

I'm proposing we do something similar for the `webapp` component. Open for discussion!