Fix hover menu z-index
The hover menu is a child of the collapsible checklist, which contains a
z-index to handle the sticky checklist headers. However, this makes the
menu be hidden under the checklists that are below. The solution
proposed here makes the n-th checklist be on top of the n-th checklist
by making `z-index(n-th checklist) = 1 + z-index((n+1)-th checklist)`.