MM-38588: Manage checklists (#920)
* Add Modal assets

* Add endpoint to create a new checklist

* Add button and dialog to create a new checklist

* Style new checklists with no tasks

* Add endpoint to rename a checklist

* Generalize DotMenu styling

* Add rename dialog component

* Add run ID prop only when the checklist is enabled

* Refactor out hover menu with rename/delete options

* Fix hover menu z-index

The hover menu is a child of the collapsible checklist, which contains a
z-index to handle the sticky checklist headers. However, this makes the
menu be hidden under the checklists that are below. The solution
proposed here makes the n-th checklist be on top of the n-th checklist
by making `z-index(n-th checklist) = 1 + z-index((n+1)-th checklist)`.

* Add an endpoint to delete a checklist

* Allow destructive buttons in GenericModal

* Move the dialogs out of the hover menu

If the dialogs are the hover menu, as soon as the window loses
focus (right-click anywhere, or focusing another window) the dialog
disappears, because the menu is no longer mounted.

* Add a delete checklist dialog

* make mocks

* Client methods to create/remove/rename checklists

* Add server E2E tests for checklist management

* Make DotMenu a bit more accessible and testable

* Add E2E tests for checklist management

* make i18n-extract

* Create a new run with every test in checklist spec

* Fix test that had a dependency on the previous one

* Modify endpoint to add checklist to be a POST

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>