Allow creating runs in an existing channel via API (#975)
* Allow creating runs in an existing channel via API

Fundamentally, we don't require a /new/ channel for a run, but currently tie the two steps together. Let's start decoupling that by allowing -- at an API level -- the ability to specify an existing channel into which to create a playbook run.

Note that I'm using Cypress E2E tests to exercise this API vs. unit tests.

* simplify from code review feedback

* test revisions (#976)

Co-authored-by: Cass C <cass.caplinger@mattermost.com>