MM-40564: Prefer native `href` links where possible (#1011)
* refactor(server): delete period from error string

golangci-lint complained when I was locally
validating my changes.

* refactor(webapp): prefer native href links

* feat: create <RHSTitleLink> as a styled <Link>

<RHSTitleText> was/is a <button> and was being used elsewhere so the
happy path forward was duplicating the styles to a React <Link> which
maintains the look & feel but gives the desired href-style links.

* refactor(webapp): <PlaybookDiv> -> <PlaybookLink>

I had initially swapped the <PlaybookDiv> for a vanilla <Link> with 2
inlined-styles. This commit refactors further by renaming the
now-defunct <PlaybookDiv> to <PlaybookLink> which styles a vanilla
<Link> so the inline styles are now also not needed.

* refactor(webapp): duplicate specificity with `&&`

* fix(webapp): `&&` -> `&&&` increase specificity

Needed to override pseudoclasses e.g. :hover and :active.

Co-authored-by: Alejandro García Montoro <alejandro.garciamontoro@gmail.com>
Co-authored-by: Mattermod <mattermod@users.noreply.github.com>