Fix infinite scroll column sorting in run list
When sorting runs differently (using a different field or setting the
opposite direction), the page needs to be reset to 0. Otherwise, the
useRunsList hook will simply add the newly fetched runs to the existing
array, potentially having duplicate elements. This, in turns, makes the
InfiniteScroll component buggy, as it ends up containing elements with
the same key value.