Webapp: Make top margin in GenericModal adjustable
The GenericModal positioning is far from great. We need to come up with
a cleaner way to center it, but the styling from bootstrap and from
webapp is interfering. This (hack-ish) solution allows the caller to
specify a number of pixels to adjust the top margin, which is enough for
now to position it vertically.