Server: Parametrize GetChannelActions with options
Until now, the GetChannelActions endpoint accepted a single parameter:
the trigger type. With the new action added, we now need a parameter to
filter by action type as well. In order to manage these options in a
more structured way, both parameters are decoded into a struct holding
all the possible options; namely: trigger and action types. This commit
adds the new parameter to the endpoint and the options struct to all the
related functions, down to the store.