Server: Adapt MessageHasBeenPosted to new logic
The new handler for the MessageHasBeenPosted hook is a bit simpler:
instead of retrieving all the (cached) playbooks, their keywords and
comparing the message against them, we now know which playbooks to
retrieve, as the action contains the pair keywords-playbookID. This
removes the need to cache the playbooks, as we don't need to look at all
of them, only at the playbooks present in the action, and only if the
keywords are present in the message.
Although the current UI only allows one keywords-playbookID pair, this
code is ready for more than one.