Webapp: Merge fetched actions with default ones
When fetching the actions, we can no longer replace a whole array of
actions from a trigger, because the fetched actions may have just a
subset of all the possible actions from the trigger. This commit handles
that case, merging the default actions for a trigger with the fetched
ones, so all default actions are shown if there are not actions in the
fetched array to overwrite them.