Webapp: Improve category selector: menu position and placeholder
- The category selector menu should be placed on the bottom of the input
box in the actions modal, while remaining on the top of the box in the
playbook action. Thus, we need to expose that prop to the callers so it
can be configurable.
- The placeholder was never shown because the value prop passed to
StyledCreatable was always a non-null object, even if props.categoryName
was null.
- The placeholder prop is also now mandatory, as having a default in the
CategorySelector component itself is not very useful. That default value
was moved to the only usage of the component in the playbook edit screen.

fixup with category selector improvements: this one makes placeholder mandatory