Webapp: Only update the actions if they've changed
To avoid making useless requests, we should only update the actions when
they have changed. This commit solves this in a primitive way: every
time onUpdateAction is called, we assume the actions have changed. This
may not be always the case, as we could have checked and unchecked an
action, leaving it in the same state, but it's a good first
approximation for those cases where nothing changes.