Server: Move UserHasJoinedChannel function to ChannelActionService
This commit simply splits the UserHasJoinedChannel function into two
methods: the old one in PlaybookRunService and a new one in
ChannelActionService. It makes no changes to the code inside the
function at all.

The goal of this commit is to isolate the meaningless diff (i.e., simply
moving code around) in a single commit so that the real changes can be
easily inspected in the following commits.

Note that this commit does not compile.