Server: Edit the message and fix the timestamp bug
After a user clicks on the "Yes, run playbook" button, the original
prompt message is edited with the copy "@user ran the PlaybookName
playbook", also removing the original buttons.

This change also fixes a bug that made the edited post look like it was
sent on January 1st 1970. This happened because a whole new post was
created, and it is fixed by simply editing the original post structure.