Server: Add promptPostID parameter to OpenCreatePlaybookRunDialog
This new parameter contains the ID of the post that was automatically
sent by the bot prompting to run a playbook, which was triggered when
certain keywords were posted in a message in the channel.

If the dialog is not spawn from such a prompt post, then this parameter
is empty.