Server: Edit the prompt post message once the run is actually created
When a playbook is run from the "Yes, run playbook" button in a prompt
post, we need to edit it in order to show that the run was actually
started. In order to do this, we need to pass the ID of the post in
the dialog state, which is then retrieved by the handler to edit the
message with the new copy once the run is created.