MM-41267: Keywords prompt playbook action (#1081)
* Server: Add IsFromPoster method to Poster

Checks whether the passed post was posted by the Playbooks bot

* Server: Parametrize GetChannelActions with options

Until now, the GetChannelActions endpoint accepted a single parameter:
the trigger type. With the new action added, we now need a parameter to
filter by action type as well. In order to manage these options in a
more structured way, both parameters are decoded into a struct holding
all the possible options; namely: trigger and action types. This commit
adds the new parameter to the endpoint and the options struct to all the
related functions, down to the store.

* Server: Add keyword trigger, run playbook action

This commit introduces the new trigger (keywords posted) and the new
action (prompt to run a playbook) into the data model, adding the needed
types, validation functions and unmarshaling from the store.

* Server: Adapt MessageHasBeenPosted to new logic

The new handler for the MessageHasBeenPosted hook is a bit simpler:
instead of retrieving all the (cached) playbooks, their keywords and
comparing the message against them, we now know which playbooks to
retrieve, as the action contains the pair keywords-playbookID. This
removes the need to cache the playbooks, as we don't need to look at all
of them, only at the playbooks present in the action, and only if the
keywords are present in the message.
Although the current UI only allows one keywords-playbookID pair, this
code is ready for more than one.

* Server: Remove code no longer used

There is no need for caching playbooks anymore, as the playbook ID is
already included in the action's payload. This commit simply removes all
the code related to that legacy structure.

* Server: Fix tests

* Server: Move MessageHasBeenPosted to ChannelActionService

This commit simply moves code, there is not much to see here: the
MessageHasBeenPosted logic is moved to ChannelActionService, greatly
simplifying PlaybookService. This also moves all the ancillary code, as
the keywords thread ignorer, and it fixes the tests.

* Server: make mocks

* Webapp: Remove keyword action from playbook

The new action is now isolated to individual channels, so we need to get
rid of it in the playbook edit and preview screens.

* Webapp: Simplify InputKeywords component

After removing the component for the keywords action in the playbook,
this component can be simplified. I also changed its name to
KeywordsSelector and generalized it for future usages if needed.

* Webapp: Make top margin in GenericModal adjustable

The GenericModal positioning is far from great. We need to come up with
a cleaner way to center it, but the styling from bootstrap and from
webapp is interfering. This (hack-ish) solution allows the caller to
specify a number of pixels to adjust the top margin, which is enough for
now to position it vertically.

* Webapp: Make titles const using FormattedMessage

Instead of using formatMessage, which is returned by the useIntl hook
we can use the FormattedMessage component, which allows to refactor the
constant definition of titles out of the main component.

HT @calebroseland.

* Webapp: Simplify fetchChannelActions

Instead of structuring the data inside fetchChannelActions, simply
return the plain array of actions, and let the callers structure it as
they need.

* Webapp: Bucket actions by trigger

Until now, the code was structuring actions inside a map
ActionType->Action. This commit changes that basic structure to a map
that is bucketed by trigger; i.e., TriggerType->(ActionType->Action).
This is meant to simplify the rendering of multiple triggers with
multiple actions inside.

* Webapp: Add keyword trigger, run playbook action

Add the UI compponents for the new trigger and action, as well as the
needed types.

* Webapp: make i18n-extract

* Webapp: disable keywords selector for regular members

I had forgotten to do this. The keywords selector in the trigger should
be disabled for regular channel members, as actions selectors are.

* Server: Track execution of keywords action

* Server: Make function to get playbook URL public

We'll need it in the api package, so this commit exposes the
GetPlaybookDetailsRelativeURL function publicly.

* Server: Edit the message and fix the timestamp bug

After a user clicks on the "Yes, run playbook" button, the original
prompt message is edited with the copy "@user ran the PlaybookName
playbook", also removing the original buttons.

This change also fixes a bug that made the edited post look like it was
sent on January 1st 1970. This happened because a whole new post was
created, and it is fixed by simply editing the original post structure.

* Server: Make linter happy

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>