Webapp: Add scrollbars around the modal contents
With the three actions configured, the modal is too high, so we need to
add scrollbars around its central part.

This commit, though, is basically a fight against the react-bootstrap
GenericModal. The idea is to implement the UI as designed:
1. There's a border on the bottom on the header and on the top of the
footer, enclosing the central content.
2. The scrollbars are close to the right side of the modal (8px away,
specifically).

For 1., the styled components ModalHeader and ModalFooter are used: they
use the :after pseudo-element to add those borders. As the GenericModal
has a padding all around the outer container, we cannot add a bottom
border to the header, as it would just span the width of the header.
Using these pseudo-elements, the line does span the whole width.

For 2., we use the renderThumbVertical and renderTrackVertical
functions. The first one sets the position of the thumb as fixed and
with right set to 8px to position it as expected. The second one is used
for the padding, so that the thumb does not touch the borders added in
1. You'll find a comment there explaining that although we pass ...style
to it, we need to manually set height, top and right to actually make
the scrollbar visible and positioned where it should be.

Finally, the styling in the first and last children of type Trigger is
set to add the desired spacing on top and bottom of all the triggers,
while letting them be hidden just below the borders added in 1. If we
add those margins to the TriggersContainer, then when scrolling, the
central contents are hidden below the top border (and over the bottom
one).