Set the ID of an action when it is saved
When saving an action, the saveChannelAction decides whether to send a
PUT or a POST request depending on the ID of the action. If it exists,
it means that the action is already in the server, and we need to update
it, so it sends a PUT. If it is not, the action does not exist in the
server yet, so it sends a POST.

Not having the ID set in the action, though, means that we can end up
with duplicated actions every time we save. This commit makes sure that
right after the action is saved, the ID is set.