MM-42437: Do not broadcast creation if broadcast is disabled (#1088)
* Do not broadcast creation if broadcast is disabled

When a playbook run is being created, and only then, we can assume that
its configuration is an exact copy of the configuration of its playbook.
While that assumption is correct, its interpretation in this piece of
code was not right. The configuration of the playbook is not only
playbook.BroadcastChannelIDs, but that value along with
playbook.BroadcastEnabled. If the broadcast is enabled in the
playbook (regardless of the value of playbook.BroadcastChannelIDs, which
may contain some non-empty value that was stored from a previous
configuration), then playbookRun.BroadcastChannelIDs is guaranteed to be
empty.

The values in the playbook run struct are always valid because, on
creation, we do check if the playbook corresponding configuration is
enabled or not, only copying the values if it is enabled.

That's why we need to use playbookRun.BroadcastChannelIDs when
broadcasting the announcement of the playbook run, and not the one in
the playbook struct.

* Test playbook with disabled broadcast but non-empty broadcast data

Check that a playbook with broadcast disabled, but with data in
broadcastChannelIds, does not broadcast the announcement nor the status
updates.

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>