MM-42700: Fix duplicated actions (#1098)
* Set the ID of an action when it is saved

When saving an action, the saveChannelAction decides whether to send a
PUT or a POST request depending on the ID of the action. If it exists,
it means that the action is already in the server, and we need to update
it, so it sends a PUT. If it is not, the action does not exist in the
server yet, so it sends a POST.

Not having the ID set in the action, though, means that we can end up
with duplicated actions every time we save. This commit makes sure that
right after the action is saved, the ID is set.

* Disallow creating more than one action with the same type

For now, we only allow one action with a specific pair (action type,
trigger type). This commit makes it so that it is impossible to create
duplicates via the API.

* Call Setup and CreateBasic just once

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>