MM-42728: fix every other chart label
The labelling of the x-axis intentionally omits every other tick so as not to clutter the UI, but this unintentionally results in that data not displaying on hover. Reference the original labels prop when rendering instead.

I haven't written a unit test for this, since it involves hovering over a graph and I don't know how to do that.

Fixes: https://mattermost.atlassian.net/browse/MM-42728