MM-38534: Remove message button on run overview
The "Message" button has been prsent on the run overview for sometime, ostensibly to allow users to directly reach out to the participants in question via a DM. I can't recall ever using this, and there are issues rendering this page with long titles.

Proposing we just remove the messaging functionality altogether as we look forward to a revamped run overview page in Q2.

Fixes: https://mattermost.atlassian.net/browse/MM-38534