MM-42509: Change initial fetch to get only in progress runs (#1113)
* Change initial fetch to get only in progress runs

* Check for finished runs when the RHS opens

Now that the store only includes in-progress runs by default, we need to
cover the case when a user manually opens the RHS to see the checklist
of a finished run.

This is solved by asking the server whether the current channel has a
run associated. We could make this request every time the user switches
channels, but that would have a negative impact in the performance of
the overall product. Instead, we only ask the server whether the channel
has a run associated *if* the RHS is open and *if* the store does not
contain a run associated to that channel. The only downside of this
solution is that there's a flickering in the rendered RHS: it first
shows the "This channel is not running any playbook" screen and when the
sever responds, it switches to the run RHS with the checklist.

* Test RHS opening on a finished run

* Remove it.only