Refactor ActionsModal
- Refactor the channel-actions-specific logic out of ActionsModal, so
that the ActionsModal component is just a wrapper of GenericModal
preparing the field for a new RunActionsModal.
- Add a new ChannelActionsModal component that uses the now generic
ActionsModal. This new component is what was called ActionsModal, and
it's the one that is shown when clicking on ChannelActions in the
channel header menu entry.
- Simplify the Action and Trigger components: they now receive only the
props the need, not the whole ChannelAction object. Also, they no longer
are the owners of setting the titles strings, this is now done by the
caller that renders the component. All in all, they're simply UI
wrappers with little logic, which is now managed by the callers.