Test saving empty actions, which is now allowed
There can be only one action of each type, so we need to create a whole
new test environment for each unit test. That's why this commit copies
the Setup and CreateBasic functions to each individual test. Otherwise,
some of the tests fail because a previous one already created an action
of the same type.