Reset actions to empty state on channel switch
When the user switches between channels, the channel actions are fetched
to populate the actions modal. If there are actions, their state is set
to the fetched data. However, if there is some missing action, nothing
happened, as we only iterated through the returned actions. This commit
changes that, to ensure the state of all actions is reset: for those
fetched and for those that are not fetched.

Add an E2E test to check this case.