Fix the modal rendering a previous step bug
Whenever the actions modal was saved, the underlying component
GenericModal was still executing onHide, even when
autoCloseOnConfirmButton was set to false. This happened because
GenericModal set the onExited prop to the passed onHide: this caused
onHide to be called after the modal had existed.

This created a race condition in ChannelActionsModal, which manages two
pieces of state to store the current state of the actions entered by the
user (let's call that ClientState) and the state of the actions stored
in the server (let's call that ServerState). This is supposed to work
like this:
1. When the modal is saved, ServerState is overwritten with whatever
ClientState had, as this is the information we are sending to the server.
2. When the modal is canceled (with the Cancel button or because the
user clicks outside of it), the ClientState is overwritter with whatever
ServerState had, as the changes are being discarded.

As onHide was incorrectly being called when onSave was executed, 2.
happened alongside with 1. In a best case scenario, this was useless, as
right after 1., ClientState and ServerState were exactly the same. In a
worst case scenario, this caused a bug, as ClientState was set to
ServerState *before* ServerState had the correct information.

This race condition was rare, and only happened on an E2E test, although
there, it was consistently failing. It was nearly impossible to
reproduce this manually.